split extension into multiple files, introduce flow mutext, don't emit stream errors as this will close the stream, use invitation topology
