+++ packages/sdk/client-services/src/packlets/invitations/invitation-host-extension.ts
@@ -2,7 +2,7 @@
 // Copyright 2022 DXOS.org
 //
 
-import { Trigger } from '@dxos/async';
+import { type Mutex, type MutexGuard, Trigger } from '@dxos/async';
 import { cancelWithContext, Context } from '@dxos/context';
 import { randomBytes, verify } from '@dxos/crypto';
 import { invariant } from '@dxos/invariant';
@@ -70,7 +70,15 @@ export class InvitationHostExtension extends RpcExtension<
    */
   public completedTrigger = new Trigger<PublicKey>();
 
-  constructor(private readonly _callbacks: InvitationHostExtensionCallbacks) {
+  /**
+   * Held to allow only one invitation flow at a time to be active.
+   */
+  private _invitationFlowLock: MutexGuard | null = null;
+
+  constructor(
+    private readonly _invitationFlowMutex: Mutex,
+    private readonly _callbacks: InvitationHostExtensionCallbacks,
+  ) {
     super({
       requested: {
         InvitationHostService: schema.getService('dxos.halo.invitations.InvitationHostService'),
@@ -109,9 +117,7 @@ export class InvitationHostExtension extends RpcExtension<
           }
           this.invitation = invitation;
 
-          log('guest introduced itself', {
-            guestProfile: profile,
-          });
+          log('guest introduced themselves', { guestProfile: profile });
           this.guestProfile = profile;
 
           this._callbacks.onStateUpdate({ ...this.invitation, state: Invitation.State.READY_FOR_AUTHENTICATION });
@@ -211,12 +217,14 @@ export class InvitationHostExtension extends RpcExtension<
     await super.onOpen(context);
 
     try {
+      this._invitationFlowLock = await cancelWithContext(this._ctx, this._invitationFlowMutex.acquire());
       await this.rpc.InvitationHostService.options({ role: Options.Role.HOST });
       await cancelWithContext(this._ctx, this._remoteOptionsTrigger.wait({ timeout: OPTIONS_TIMEOUT }));
       if (this._remoteOptions?.role !== Options.Role.GUEST) {
         throw new InvalidInvitationExtensionRoleError(undefined, {
           expected: Options.Role.GUEST,
           remoteOptions: this._remoteOptions,
+          remotePeerId: context.remotePeerId,
         });
       }
 
@@ -227,83 +235,8 @@ export class InvitationHostExtension extends RpcExtension<
   }
 
   override async onClose() {
-    await this._ctx.dispose();
-  }
-}
-
-type InvitationGuestExtensionCallbacks = {
-  // Deliberately not async to not block the extensions opening.
-  onOpen: (ctx: Context) => void;
-  onError: (error: Error) => void;
-};
-
-/**
- * Guest's side for a connection to a concrete peer in p2p network during invitation.
- */
-export class InvitationGuestExtension extends RpcExtension<
-  { InvitationHostService: InvitationHostService },
-  { InvitationHostService: InvitationHostService }
-> {
-  private _ctx = new Context();
-  private _remoteOptions?: Options;
-  private _remoteOptionsTrigger = new Trigger();
-
-  constructor(private readonly _callbacks: InvitationGuestExtensionCallbacks) {
-    super({
-      requested: {
-        InvitationHostService: schema.getService('dxos.halo.invitations.InvitationHostService'),
-      },
-      exposed: {
-        InvitationHostService: schema.getService('dxos.halo.invitations.InvitationHostService'),
-      },
-    });
-  }
-
-  protected override async getHandlers(): Promise<{ InvitationHostService: InvitationHostService }> {
-    return {
-      InvitationHostService: {
-        options: async (options) => {
-          invariant(!this._remoteOptions, 'Remote options already set.');
-          this._remoteOptions = options;
-          this._remoteOptionsTrigger.wake();
-        },
-        introduce: () => {
-          throw new Error('Method not allowed.');
-        },
-        authenticate: () => {
-          throw new Error('Method not allowed.');
-        },
-        admit: () => {
-          throw new Error('Method not allowed.');
-        },
-      },
-    };
-  }
-
-  override async onOpen(context: ExtensionContext) {
-    await super.onOpen(context);
-
-    try {
-      log('begin options');
-      await cancelWithContext(this._ctx, this.rpc.InvitationHostService.options({ role: Options.Role.GUEST }));
-      await cancelWithContext(this._ctx, this._remoteOptionsTrigger.wait({ timeout: OPTIONS_TIMEOUT }));
-      log('end options');
-      if (this._remoteOptions?.role !== Options.Role.HOST) {
-        throw new InvalidInvitationExtensionRoleError(undefined, {
-          expected: Options.Role.HOST,
-          remoteOptions: this._remoteOptions,
-        });
-      }
-
-      this._callbacks.onOpen(this._ctx);
-    } catch (err: any) {
-      log('openError', err);
-      this._callbacks.onError(err);
-    }
-  }
-
-  override async onClose() {
-    log('onClose');
+    this._invitationFlowLock?.release();
+    this._invitationFlowLock = null;
     await this._ctx.dispose();
   }
 }
